added filter ui for store
