+++ baseModels.json
@@ -1,0 +1,22 @@
+{
+  "baseModels": [
+    "Other",
+    "PixArt a",
+    "Playground v2",
+    "SD 1.4",
+    "SD 1.5",
+    "SD 1.5 LCM",
+    "SD 2.0",
+    "SD 2.0 768",
+    "SD 2.1",
+    "SD 2.1 768",
+    "SD 2.1 Unclip",
+    "SDXL 0.9",
+    "SDXL 1.0",
+    "SDXL 1.0 LCM",
+    "SDXL Distilled",
+    "SDXL Turbo",
+    "SVD",
+    "SVD XT"
+  ]
+}
